refactor shade
