+++ fe/fe-common/pom.xml
@@ -108,6 +108,10 @@ under the License.
             </exclusions>
             <scope>provided</scope>
         </dependency>
+        <dependency>
+            <groupId>io.trino</groupId>
+            <artifactId>trino-main</artifactId>
+        </dependency>
     </dependencies>
     <build>
         <finalName>doris-fe-common</finalName>
