+++ fe/fe-common/src/main/java/org/apache/doris/trinoconnector/TrinoColumnMetadata.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.doris.datasource.trinoconnector.shade;
+package org.apache.doris.trinoconnector;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
