+++ fe/fe-common/src/main/java/org/apache/doris/trinoconnector/TrinoConnectorPluginManager.java
@@ -14,11 +14,13 @@
 // KIND, either express or implied.  See the License for the
 // specific language governing permissions and limitations
 // under the License.
+// copied from https://github.com/trinodb/trino/blob/master/core/trino-main/src/main/java/io/trino/server/PluginManager.java
 
-package org.apache.doris.trinoconnector.shade;
+package org.apache.doris.trinoconnector;
+
+import org.apache.doris.common.EnvUtils;
 
 import com.google.common.base.Preconditions;
-import com.google.common.base.Strings;
 import com.google.common.collect.ImmutableList;
 import io.trino.connector.ConnectorName;
 import io.trino.metadata.HandleResolver;
@@ -44,8 +46,8 @@ import java.util.function.Supplier;
 import javax.annotation.concurrent.ThreadSafe;
 
 @ThreadSafe
-public class TrinoConnectorPluginManager implements PluginInstaller  {
-    public static final String trinoConnectorPluginsDir = getDorisHome() + "/lib/connectors";
+public class TrinoConnectorPluginManager implements PluginInstaller {
+    public static final String trinoConnectorPluginsDir = EnvUtils.getDorisHome() + "/lib/connectors";
     private static final ImmutableList<String> SPI_PACKAGES = ImmutableList.<String>builder()
             .add("io.trino.spi.")
             .add("com.fasterxml.jackson.annotation.")
@@ -89,9 +91,9 @@ public class TrinoConnectorPluginManager implements PluginInstaller  {
 
         PluginClassLoader pluginClassLoader = createClassLoader.get();
 
-        LOG.info("Classpath for plugin:");
+        LOG.debug("Classpath for plugin:");
         for (URL url : pluginClassLoader.getURLs()) {
-            LOG.info("    {}", url.getPath());
+            LOG.debug("    {}", url.getPath());
         }
 
         handleResolver.registerClassLoader(pluginClassLoader);
@@ -133,14 +135,6 @@ public class TrinoConnectorPluginManager implements PluginInstaller  {
     public static PluginClassLoader createClassLoader(String pluginName, List<URL> urls) {
         ClassLoader parent = TrinoConnectorPluginManager.class.getClassLoader();
         return new PluginClassLoader(pluginName, urls, parent, SPI_PACKAGES);
-    }
-
-    private static String getDorisHome() {
-        String dorisHome = System.getenv("DORIS_HOME");
-        if (Strings.isNullOrEmpty(dorisHome)) {
-            dorisHome = "${Env.DORIS_HOME}";
-        }
-        return dorisHome;
     }
 
     public ConcurrentMap<ConnectorName, ConnectorFactory> getConnectorFactories() {
