+++ fe/fe-core/src/main/java/org/apache/doris/catalog/Env.java
@@ -135,7 +135,7 @@ import org.apache.doris.datasource.es.EsExternalCatalog;
 import org.apache.doris.datasource.es.EsRepository;
 import org.apache.doris.datasource.hive.HiveTransactionMgr;
 import org.apache.doris.datasource.hive.event.MetastoreEventsProcessor;
-import org.apache.doris.datasource.trinoconnector.shade.TrinoConnectorPluginManager;
+import org.apache.doris.trinoconnector.TrinoConnectorPluginManager;
 import org.apache.doris.deploy.DeployManager;
 import org.apache.doris.deploy.impl.AmbariDeployManager;
 import org.apache.doris.deploy.impl.K8sDeployManager;
