+++ fe/fe-core/src/main/java/org/apache/doris/datasource/trinoconnector/source/TrinoConnectorScanNode.java
@@ -29,8 +29,8 @@ import org.apache.doris.common.AnalysisException;
 import org.apache.doris.common.DdlException;
 import org.apache.doris.common.MetaNotFoundException;
 import org.apache.doris.common.UserException;
-import org.apache.doris.datasource.trinoconnector.shade.TrinoColumnMetadata;
-import org.apache.doris.datasource.trinoconnector.shade.TrinoConnectorPluginManager;
+import org.apache.doris.trinoconnector.TrinoColumnMetadata;
+import org.apache.doris.trinoconnector.TrinoConnectorPluginManager;
 import org.apache.doris.nereids.glue.translator.PlanTranslatorContext;
 import org.apache.doris.planner.PlanNodeId;
 import org.apache.doris.spi.Split;
